==> 1-tutorial-video-extract.ipynb <==
# https://www.analyticsvidhya.com/blog/2019/09/step-by-step-deep-learning-tutorial-video-classification-python/
# Copy video data from https://www.crcv.ucf.edu/data/UCF101/UCF101.rar
# Expand in test/data get folder UCF-101 which many videos.

import cv2     # for capturing videos
import math   # for mathematical operations
import matplotlib.pyplot as plt    # for plotting the images
%matplotlib inline
import pandas as pd
from keras.preprocessing import image   # for preprocessing the images
import numpy as np    # for mathematical operations
from keras.utils import np_utils
from skimage.transform import resize   # for resizing images
from sklearn.model_selection import train_test_split
from glob import glob
from tqdm import tqdm

f = open("test/data/ucfTrainTestlist/testlist01.txt", "r")
temp = f.read()
videos = temp.split('\n')
train = pd.DataFrame()
train['video_name'] = videos
train = train[:-1]
train.head()

train_video_tag = []
for i in range(train.shape[0]):
    train_video_tag.append(train['video_name'][i].split('/')[0])
    
train['tag'] = train_video_tag
train.head()

def expand(indir, train, outdir):
    
    import os
    os.makedirs(outdir, exist_ok=True)

    print("Loop over {}".format(train.shape[0]))
    for i in tqdm(range(train.shape[0])):
        count = 0
        videoFile = train['video_name'][i]
        inpath = os.path.join(indir, videoFile)
        if not os.path.isfile(inpath):
            raise Exception("Path does not exist {}".format(inpath))
            
        cap = cv2.VideoCapture(inpath)   # capturing the video from the given path
        try:
            frameRate = cap.get(5) #frame rate
            x=1
            while(cap.isOpened()):
                frameId = cap.get(1) #current frame number
                ret, frame = cap.read()
                if (ret != True):
                    break
                if (frameId % math.floor(frameRate) == 0):
                    # storing the frames in a new folder named train_1
                    filename = videoFile.split('/')[1].split(' ')[0] +"_frame%d.jpg" % count;count+=1
                    outpath = os.path.join(outdir, filename)
                    if os.path.isfile(outpath):
                        raise Exception("File {} already exists!".format(outpath))
                    cv2.imwrite(outpath, frame)
                    if i % 1000 == 0:
                        print(outpath)
        finally:
            cap.release()
            

expand("test/data/UCF-101", train, "scratch/train_01")

def save_frame(indir, pattern="*.jpg", sort=False):
    
    import re
    p = re.compile("v_([a-zA-Z]+)_(.*\.avi_frame\d+).jpg")
    
    images = glob("scratch/train_01/"+pattern)
    train_image = []
    train_class = []
    for i in tqdm(range(len(images))):
        
        filename = os.path.basename(images[i])
        m = p.match(filename)
        
        # creating the image name
        train_image.append(m.group(1)+"_"+m.group(2))
        
        # creating the class of image
        train_class.append(m.group(1))

    # storing the images and their class in a dataframe
    train_data = pd.DataFrame()
    train_data['image'] = train_image
    train_data['class'] = train_class

    if sort:
        train_data = train_data.sort_values(by=['image'])
    
    return train_data

td = save_frame("scratch/train_01", sort=True)

# converting the dataframe into csv file 
td.to_csv('train_new.csv',header=True, index=False)

td.head()

import os
print(len([name for name in os.listdir('scratch/train_01') if os.path.isfile('scratch/train_01/'+name)]))
