==> 5-train-sport-with-k-average.ipynb <==
# See https://www.pyimagesearch.com/2019/07/15/video-classification-with-keras-and-deep-learning/
from tensorflow.keras.preprocessing.image import ImageDataGenerator
from tensorflow.keras.layers import AveragePooling2D
from tensorflow.keras.applications import ResNet50
from tensorflow.keras.layers import Dropout
from tensorflow.keras.layers import Flatten
from tensorflow.keras.layers import Dense
from tensorflow.keras.layers import Input
from tensorflow.keras.models import Model
from tensorflow.keras.optimizers import SGD
from sklearn.preprocessing import LabelBinarizer
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from imutils import paths
import matplotlib.pyplot as plt
import numpy as np
import argparse
import pickle
import cv2
import os
import pandas as pd
from tqdm import tqdm

def get_labels(pardir='test/data/sports/data', sort=True):
    
    from os.path import isdir, basename, join, isfile
    files = [f for f in os.listdir(pardir) if isdir(join(pardir, f))]

    train_image = []
    train_class = []

    for idir in files:
        clazz = basename(idir)
        if clazz.startswith("."):
            continue
        
        images = os.listdir(pardir+"/"+clazz)
        for i in tqdm(range(len(images))):
            
            image_name = images[i] 
            lc = image_name.lower()
            if not lc.endswith(".jpg") and not lc.endswith(".jpeg") and not lc.endswith(".png"):
                continue
            
            if isdir(join(pardir, clazz, image_name)):
                continue
                
            if lc.endswith(".ipynb_checkpoints"):
                continue
            
            train_image.append(clazz+"/"+image_name)
            train_class.append(clazz)
        
    train_data = pd.DataFrame()
    train_data['image'] = train_image
    train_data['class'] = train_class

    if sort:
        train_data = train_data.sort_values(by=['class'])
        
    return train_data

train = get_labels()

def images_and_labels(pdir='test/data/sports/data', train=train):
    
    # creating an empty list
    images = []
    labels = []

    # for loop to read and store frames
    for i in tqdm(range(train.shape[0])):
        # loading the image and keeping the target size as (500,300,3)
        # The images are variable size
        impath = pdir+'/'+train['image'][i]
        if i % 1000 == 0:
            print(impath)
        try:
            image = cv2.imread(impath)
            image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            image = cv2.resize(image, (224, 224))
            
            # appending the image to the train_image list
            images.append(image)
            labels.append(train['class'][i])
        except:
            raise Exception("Problem with {}".format(impath))
    return images, labels

data, labels = images_and_labels()

data = np.array(data)
labels = np.array(labels)

# perform one-hot encoding on the labels
lb = LabelBinarizer()
labels = lb.fit_transform(labels)

# partition the data into training and testing splits using 75% of
# the data for training and the remaining 25% for testing
(trainX, testX, trainY, testY) = train_test_split(data, labels,
	test_size=0.25, stratify=labels, random_state=42)

# initialize the training data augmentation object
trainAug = ImageDataGenerator(
	rotation_range=30,
	zoom_range=0.15,
	width_shift_range=0.2,
	height_shift_range=0.2,
	shear_range=0.15,
	horizontal_flip=True,
	fill_mode="nearest")

# initialize the validation/testing data augmentation object (which
# we'll be adding mean subtraction to)
valAug = ImageDataGenerator()

# define the ImageNet mean subtraction (in RGB order) and set the
# the mean subtraction value for each of the data augmentation
# objects
mean = np.array([123.68, 116.779, 103.939], dtype="float32")
trainAug.mean = mean
valAug.mean = mean

# load the ResNet-50 network, ensuring the head FC layer sets are left
# off
baseModel = ResNet50(weights="imagenet", include_top=False,
	input_tensor=Input(shape=(224, 224, 3)))


# construct the head of the model that will be placed on top of the
# the base model
headModel = baseModel.output
headModel = AveragePooling2D(pool_size=(7, 7))(headModel)
headModel = Flatten(name="flatten")(headModel)
headModel = Dense(512, activation="relu")(headModel)
headModel = Dropout(0.5)(headModel)
headModel = Dense(len(lb.classes_), activation="softmax")(headModel)

# place the head FC model on top of the base model (this will become
# the actual model we will train)
model = Model(inputs=baseModel.input, outputs=headModel)

# loop over all layers in the base model and freeze them so they will
# *not* be updated during the training process
for layer in baseModel.layers:
    layer.trainable = False

def compile(epochs=50, model=model) :
    # compile our model (this needs to be done after our setting our
    # layers to being non-trainable)
    print("[INFO] compiling model...")
    opt = SGD(lr=1e-4, momentum=0.9, decay=1e-4 / epochs)
    model.compile(loss="categorical_crossentropy", optimizer=opt,
        metrics=["accuracy"])

compile()


def train(epochs=25, model=model):
    print("[INFO] training head...")
    from keras.callbacks import ModelCheckpoint
    mcp_save = ModelCheckpoint('sports_weight.h5', save_best_only=True, monitor='val_loss', mode='min')

    H = model.fit(
        x=trainAug.flow(trainX, trainY, batch_size=32),
        steps_per_epoch=len(trainX) // 32,
        validation_data=valAug.flow(testX, testY),
        validation_steps=len(testX) // 32,
        epochs=epochs,
        callbacks=[mcp_save])

train()

predictions = model.predict(x=testX.astype("float32"), batch_size=32)
print(classification_report(testY.argmax(axis=1), predictions.argmax(axis=1), target_names=lb.classes_))

model.save("sports_model.h5", save_format="h5")

# serialize the label binarizer to disk
f = open("sports_labels.pkl", "wb")
f.write(pickle.dumps(lb))
f.close()
