==> 6-predict-sport-with-k-average.ipynb <==
from tensorflow.keras.models import load_model
from collections import deque
import numpy as np
import argparse
import pickle
import cv2

model = load_model("sports_model.h5")
lb = pickle.loads(open("sports_labels.pkl", "rb").read())

def predict_frames(vid):
    
    mean = np.array([123.68, 116.779, 103.939][::1], dtype="float32")
    Q = deque(maxlen=128)
    
    # Get a sport video
    vs = cv2.VideoCapture(vid)

    iframe = 0
    grabbed = True
    
    print("File {}".format(vid))
    while grabbed:
        # read the next frame from the file
        (grabbed, frame) = vs.read()
        if not grabbed:
            break

        output = frame.copy()
        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
        frame = cv2.resize(frame, (224, 224)).astype("float32")
        frame -= mean

        preds = model.predict(np.expand_dims(frame, axis=0))[0]
        ipred = np.argmax(preds)
        label = lb.classes_[ipred]
        
        Q.append(preds)
        smooth = np.array(Q).mean(axis=0)
        ipred = np.argmax(smooth)
        smthLbl = lb.classes_[ipred]
        
        print("frame{} predicts {} smoothed this is {}".format(iframe, label, smthLbl))
 
        
        iframe += 1

predict_frames(vid = "test/data/example_clips/lifting.mp4")

predict_frames(vid = "test/data/example_clips/soccer.mp4")

predict_frames(vid = "test/data/example_clips/tennis.mp4")
